Moves form helper CourseSelectionStepHelper
