+++ app/forms/candidate_interface/continuous_applications/course_site_step.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     class CourseSiteStep < DfE::Wizard::Step
-      include Concerns::CourseSelectionStepHelper
+      include CandidateInterface::Concerns::CourseSelectionStepHelper
       attr_accessor :provider_id, :course_id, :study_mode, :course_option_id
       validates :course_option_id, presence: true
