+++ app/forms/candidate_interface/continuous_applications/which_course_are_you_applying_to_step.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     class WhichCourseAreYouApplyingToStep < DfE::Wizard::Step
-      include Concerns::CourseSelectionStepHelper
+      include CandidateInterface::Concerns::CourseSelectionStepHelper
       attr_accessor :provider_id, :course_id
       validates :provider_id, :course_id, presence: true
